refactor(RunMigrations): add migration to `RunMigrations`
Also adjust the order of operations on the movement and migration for safety (migrate before moving the repo).

Ensure we don't remove anything we shouldn't if the ipfsRepoPath and newRepoPath are the same. We still want to ensure the repo has been migrated, so we should still go through the motions if the old repo path and new repo path are the same.