refactor: muxfs uses the first cafs.Filesystem as a default write destination
A number of adjustments to make working with qfs easier higher up in the stack.

The most important is introducing a "default write destination", which for now
is just the first filesystem a muxfs is given that supports the cafs.Filstore
interface. In the future we can add a field to qfs.Config for manual control
of defaults.

Other notable changes:

* add `Type()` method to Filesystems, drop cafs.PathPrefix, use this instead of
  requiring a type string argument when adding a filesystem
* adjust muxfs filesystem getters & setters, use setter in New constructor,
  getter now only returns nil if the requested filesystem doesn't exist
* drop `CAFSStoreFromIPFS` for `DefaultWriteFS`, which may or may not be an IPFS
  fielsystem. For IPFS-specific behaviour, use `muxfs.Filesystem(qipfs.FilestoreType)`
  to get the qipfs filesystem